# spec/openapi.yaml
openapi: 3.1.0
info:
  title: ADP API
  version: 0.3.0
  description: >
    Native and compat-plane HTTP surface for ADP — see docs/pragmatic_mvp.md
    §2.4 for what is in scope and §2.5 for what is deliberately cut.

    Two planes share this file. `/api/v3` is the GitHub-compatible surface, kept
    faithful enough that unmodified `gh` works against it. `/api/adp` is the
    native plane, which has no GitHub analogue: workspaces, candidate sets,
    sessions and checkpoints, the operation log, and evidence bundles.

    Completeness is enforced, not aspirational: server/src/spec-coverage.test.ts
    enumerates the server's real route table and fails if this file and the
    implementation describe different APIs in either direction. Deliberate
    omissions are the git smart-HTTP wire protocol (specified by git itself and
    proxied verbatim), the loopback-only receive-path hook endpoints, and the
    inbound GitHub webhook receiver, whose payload contract is GitHub's.

    `info.version` above is the contract version, and every response carries it
    in an `ADP-API-Version` header — including 401s and 404s, so a generated
    client can pin before it authenticates. The git smart-HTTP routes are the
    exception: they are proxied verbatim and are not part of this contract.
    See docs/api-compatibility.md for what a version bump promises.
servers:
  - url: /
paths:
...
  /api/v3/repos/{owner}/{repo}/commits:
    parameters:
      - $ref: "#/components/parameters/owner"
      - $ref: "#/components/parameters/repo"
    get:
      x-required-scope: "repo:read"
      summary: Commit list, newest first. `?sha=` selects the starting ref.
      security:
        - bearerAuth: []
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "200":
          description: Commits
  /api/v3/repos/{owner}/{repo}/commits/{sha}:
    parameters:
      - $ref: "#/components/parameters/owner"
      - $ref: "#/components/parameters/repo"
      - $ref: "#/components/parameters/sha"
    get:
      x-required-scope: "repo:read"
      summary: A single commit.
      security:
        - bearerAuth: []
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "200":
          description: Commit
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not found
...
  /api/v3/repos/{owner}/{repo}/git/refs:
    parameters:
      - $ref: "#/components/parameters/owner"
      - $ref: "#/components/parameters/repo"
    get:
      x-required-scope: "repo:read"
      summary: All refs. Paginated with `per_page`/`page`.
      security:
        - bearerAuth: []
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "200":
          description: Refs
    post:
      x-required-scope: "repo:write"
      summary: Create a ref.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ref, sha]
              properties:
                ref:
                  type: string
                sha:
                  type: string
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "201":
          description: Created
  /api/v3/repos/{owner}/{repo}/git/refs/{path}:
    parameters:
      - $ref: "#/components/parameters/owner"
      - $ref: "#/components/parameters/repo"
      - $ref: "#/components/parameters/path"
    get:
      x-required-scope: "repo:read"
      summary: >
        Refs matching a prefix — GitHub's own shape, so this returns an **array**
        even when the prefix names exactly one ref.
      security:
        - bearerAuth: []
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "200":
          description: Matching refs
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: No ref matches the prefix
  /api/v3/repos/{owner}/{repo}/git/refs/heads/{branch}:
    parameters:
      - $ref: "#/components/parameters/owner"
      - $ref: "#/components/parameters/repo"
      - name: branch
        in: path
        required: true
        schema:
          type: string
    delete:
      x-required-scope: "repo:write"
      summary: Delete a branch ref.
      security:
        - bearerAuth: []
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "204":
          description: Deleted
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not found
  /api/v3/repos/{owner}/{repo}/git/blobs:
    parameters:
      - $ref: "#/components/parameters/owner"
      - $ref: "#/components/parameters/repo"
    post:
      x-required-scope: "repo:write"
      summary: Create a blob.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                encoding:
                  type: string
                  enum: [utf-8, base64]
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "201":
          description: Created
  /api/v3/repos/{owner}/{repo}/git/blobs/{sha}:
    parameters:
      - $ref: "#/components/parameters/owner"
      - $ref: "#/components/parameters/repo"
      - $ref: "#/components/parameters/sha"
    get:
      x-required-scope: "repo:read"
      summary: Read a blob.
      security:
        - bearerAuth: []
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "200":
          description: Blob
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not found
  /api/v3/repos/{owner}/{repo}/git/trees:
    parameters:
      - $ref: "#/components/parameters/owner"
      - $ref: "#/components/parameters/repo"
    post:
      x-required-scope: "repo:write"
      summary: >
        Create a tree. `base_tree` accepts any commit-ish, not only a tree sha —
        the server resolves it with `git ls-tree`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tree]
              properties:
                base_tree:
                  type: string
                tree:
                  type: array
                  items:
                    type: object
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "201":
          description: Created
  /api/v3/repos/{owner}/{repo}/git/trees/{sha}:
    parameters:
      - $ref: "#/components/parameters/owner"
      - $ref: "#/components/parameters/repo"
      - $ref: "#/components/parameters/sha"
    get:
      x-required-scope: "repo:read"
      summary: Read a tree.
      security:
        - bearerAuth: []
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "200":
          description: Tree
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not found
  /api/v3/repos/{owner}/{repo}/git/commits:
    parameters:
      - $ref: "#/components/parameters/owner"
      - $ref: "#/components/parameters/repo"
    post:
      x-required-scope: "repo:write"
      summary: Create a commit object.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message, tree, parents]
              properties:
                message:
                  type: string
                tree:
                  type: string
                parents:
                  type: array
                  items:
                    type: string
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "201":
          description: Created
  /api/v3/repos/{owner}/{repo}/git/commits/{sha}:
    parameters:
      - $ref: "#/components/parameters/owner"
      - $ref: "#/components/parameters/repo"
      - $ref: "#/components/parameters/sha"
    get:
      x-required-scope: "repo:read"
      summary: Read a commit object.
      security:
        - bearerAuth: []
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "200":
          description: Commit
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not found
...
components:
  parameters:
    owner:
      name: owner
      in: path
      required: true
      schema:
        type: string
    repo:
      name: repo
      in: path
      required: true
      schema:
        type: string
...
    sha:
      name: sha
      in: path
      required: true
      schema:
        type: string
    path:
      name: path
      in: path
      required: true
      schema:
        type: string
      description: Matches the rest of the URL, including slashes.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Error:
      type: object
      required: [message]
      description: >
        #97: the one error shape every 4xx carries. `errors` appears on
        validation failures (422) and is a deliberate projection of the
        server's internal validator output — path, message, and a stable
        code — never the validator's raw issue objects, whose shape belongs
        to a library this contract must outlive.
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            required: [path, message, code]
            properties:
              path:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              message:
                type: string
              code:
                type: string
...